+++ client-app/core/api/graphql/settings/index.ts
@@ -1,1 +1,1 @@
-export * from "./queries/getStoreSettings";
+export * from "./queries/getStore";
