+++ client-app/core/api/graphql/settings/queries/getStore/getStore.graphql
