+++ client-app/core/api/graphql/settings/queries/getStore/index.ts
